requirements_streamlit.txt
streamlit>=1.28.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
numpy>=1.21.0
pathlib
